# MAEBACKEND/Endpoints/openapi_specification.yaml
openapi: 3.0.3
info:
  title: EasyRent API - Sistema Completo de Marketplace Inmobiliario
  description: |
    API completa para el marketplace inmobiliario EasyRent con todas las funcionalidades:
    - 🔐 **Autenticación**: JWT + API Keys (Compatible con base de datos PostgreSQL)
    - 👥 **Gestión de usuarios y agencias**
    - 🏠 **Listings y multimedia**
    - 🔍 **Búsqueda avanzada y filtros**
    - 💎 **Sistema de planes y facturación**
    - 📊 **Analytics y reportes**
    - ✅ **Verificación y moderación**
    - 🔗 **Integraciones externas**
    
    ## Tecnologías
    - **Framework**: FastAPI + Python 3.11+
    - **Base de datos**: PostgreSQL 17
    - **Cache**: Redis
    - **Message Broker**: Apache Kafka
    - **Autenticación**: JWT (Bearer Token)
    
    ## Notas de Implementación
    - **Autenticación Firebase**: Sistema de autenticación completo con Firebase Authentication
    - **Flujo de autenticación**: Frontend → Firebase → Backend verification → JWT tokens
    - **Sin manejo de passwords**: Toda la gestión de credenciales delegada a Firebase
    - **Soporte híbrido**: Funciona con o sin Firebase configurado (modo desarrollo)
    - **Schemas actualizados**: Los modelos de datos coinciden con la implementación real en FastAPI
    
    ## Autenticación Firebase
    Para autenticarse, el cliente debe:
    1. Autenticarse con Firebase (frontend)
    2. Obtener el Firebase ID token
    3. Enviar el token al endpoint `/auth/login`
...
    
  version: 1.0.0
  contact:
...
servers:
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /agencies/{agency_id}/verification:
    post:
      tags: [Agencies]
      summary: Solicitar verificación de agencia
      description: Solicita la verificación oficial de la agencia con documentos
      operationId: request_verification
      parameters:
        - name: agency_id
          in: path
          required: true
          schema:
            type: string
            example: "123e4567-e89b-12d3-a456-426614174000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerificationRequest'
      responses:
        '200':
          description: Solicitud de verificación enviada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Documentos inválidos o faltantes
        '401':
          description: No autorizado
        '404':
          description: Agencia no encontrada
        '409':
          description: La agencia ya está verificada o tiene una solicitud pendiente
        '500':
          description: Error interno del servidor

  # ===========================================
  # LISTINGS MANAGEMENT
  # ===========================================
...
  /verifications:
    get:
      tags: [Verification]
      summary: Obtener verificaciones (moderador)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
        - name: type
          in: query
          schema:
            type: string
            enum: [listing, agency, user]
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Lista de verificaciones

    post:
      tags: [Verification]
      summary: Solicitar verificación
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [target_type, target_id]
              properties:
                target_type:
                  type: string
                  enum: [listing, agency, user]
                target_id:
                  type: string
                  format: uuid
                documents:
                  type: array
                  items:
                    type: string
                    format: uuid
                notes:
                  type: string
      responses:
        '201':
          description: Verificación solicitada
  /verifications/{verification_id}:
    get:
      tags: [Verification]
      summary: Obtener verificación por ID
      parameters:
        - name: verification_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Detalles de la verificación

    put:
      tags: [Verification]
      summary: Procesar verificación (moderador)
      parameters:
        - name: verification_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [approved, rejected]
                notes:
                  type: string
                requirements:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Verificación procesada
  /moderation/queue:
    get:
      tags: [Verification]
      summary: Cola de moderación
      parameters:
        - name: priority
          in: query
          schema:
            type: string
            enum: [high, medium, low]
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Cola de moderación
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Autenticación usando JWT Token.
        
        **Formato**: `Authorization: Bearer <token>`
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Autenticación usando API Key para integraciones.
        
        **Formato**: `X-API-Key: <your-api-key>`
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Autenticación usando JWT Token.
        
        **Formato**: `Authorization: Bearer <token>`
        
        **Ejemplo**: `Authorization: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9...`
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Autenticación usando API Key para integraciones.
        
        **Formato**: `X-API-Key: <your-api-key>`
  parameters:
    PageParam:
      name: page
      in: query
      description: Número de página (empezando en 1)
      schema:
        type: integer
        minimum: 1
        default: 1
        example: 1
    LimitParam:
      name: limit
      in: query
      description: Número de elementos por página
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
        example: 20
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          description: Mensaje de respuesta
          example: "Operación completada exitosamente"

    # ===========================================
    # AGENCY SCHEMAS
    # ===========================================
...
    VerificationRequest:
      type: object
      required: [documents]
      properties:
        documents:
          type: array
          items:
            type: string
            format: uri
          minItems: 1
          description: Lista de URLs de documentos para verificación
          example: [
            "https://storage.easyrent.pe/docs/ruc-certificate.pdf",
            "https://storage.easyrent.pe/docs/license.pdf"
          ]

    # ===========================================
    # LISTING SCHEMAS
    # ===========================================
...